requirements.txt
numpy==1.24.2
PyYAML==6.0
PyYAML==6.0
tqdm==4.65.0
tsplib95==0.7.1
typer==0.9.0
typing_extensions==4.5.0
